requirements.txt
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==4.4.2
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.98.0
filelock==3.12.2
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.2
lazy_loader==0.2
llm-client==0.6.1
llvmlite==0.40.1
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.3
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.3
openai==0.27.8
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
opencv-python==4.7.0.72
orjson==3.9.1
packaging==23.1
Pillow==9.5.0
proglog==0.1.10
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pyparsing==3.1.0
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
pytube @ git+https://github.com/nficano/pytube@a32fff39058a6f7e5e59ecd06a7467b71197ce35
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
semantic-version==2.10.0
setuptools-rust==1.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tifffile==2023.4.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.65.0
transliterate==1.10.2
typing_extensions==4.6.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2